send subaccount websocket msg to ender
